Reorganize test jsx file
